requirement.txt
opencv-python
easyocr
mysql-connector-python
pandas
python-dotenv
tk
